feat: move tests from DetailedHikeRoute to RouteUtils test
