+++ app/src/main/java/ch/hikemate/app/model/route/DetailedHikeRoute.kt
@@ -43,7 +43,7 @@ data class DetailedHikeRoute(
 
       val totalDistance = RouteUtils.computeTotalDistance(hikeRoute.ways)
       val elevationGain = runBlocking {
-        RouteUtils.computeElevationGain(hikeRoute.ways, hikeRoute.id, elevationService)
+        RouteUtils.getElevationGain(hikeRoute.ways, hikeRoute.id, elevationService)
       }
       val estimatedTime = RouteUtils.estimateTime(totalDistance, elevationGain)
       val difficulty = RouteUtils.determineDifficulty(totalDistance, elevationGain)
